+++ src/modules/Terra/octrees/Chunk.cs
@@ -5,7 +5,7 @@ public struct Chunk{
     public uint y{get; set;}
     public uint z{get; set;}
 
-    public Memory<byte> voxels{get; set;}
+    public Memory<uint> voxels{get; set;}
 
     public bool isEmpty{get; set;}
 }
